+++ lib/src/pigeon/age_verification_api.g.dart
@@ -442,22 +442,4 @@ class AgeVerificationApi {
     );
     return pigeonVar_replyValue! as AgeVerificationResult;
   }
-
-  Future<void> dispose() async {
-    final pigeonVar_channelName =
-        'dev.flutter.pigeon.age_verification.AgeVerificationApi.dispose$pigeonVar_messageChannelSuffix';
-    final pigeonVar_channel = BasicMessageChannel<Object?>(
-      pigeonVar_channelName,
-      pigeonChannelCodec,
-      binaryMessenger: pigeonVar_binaryMessenger,
-    );
-    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(null);
-    final pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
-
-    _extractReplyValueOrThrow(
-      pigeonVar_replyList,
-      pigeonVar_channelName,
-      isNullValid: true,
-    );
-  }
 }